server-requirements.txt
# Pacman Sync Utility Server - Python Dependencies

# Web framework and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database support
asyncpg>=0.29.0          # PostgreSQL async driver
aiosqlite>=0.19.0        # SQLite async driver

# HTTP client for external requests
aiohttp>=3.9.0
httpx>=0.25.0

# Data validation and serialization
pydantic>=2.5.0

# Configuration and environment
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Logging and monitoring
structlog>=23.2.0

# Date and time utilities
python-dateutil>=2.8.2

# JSON Web Tokens for authentication
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0

# Input validation and sanitization
bleach>=5.0.0

# Background task processing
celery>=5.3.0            # Optional for async task processing
redis>=5.0.0             # Optional for task queue backend

# Development and testing dependencies (will be installed in dev stage only)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-mock>=3.12.0
# black>=23.0.0
# flake8>=6.1.0
# mypy>=1.7.0